+++ extract_msg/utils.py
@@ -14,6 +14,7 @@ import pathlib
 import struct
 # Not actually sure if this needs to be here for the logging, so just in case.
 import sys
+import zipfile
 
 import tzlocal
 
@@ -63,6 +64,8 @@ def bitwiseAdjust(inp : int, mask : int) -> int:
     will be done automatically.
 
     :param mask: MUST be greater than 0.
+
+    :raises ValueError: if the mask is not greater than 0.
     """
     if mask < 1:
         raise ValueError('Mask MUST be greater than 0')
@@ -72,6 +75,8 @@ def bitwiseAdjustedAnd(inp : int, mask : int) -> int:
     """
     Preforms the bitwise AND operation between :param inp: and :param mask: and
     adjusts the results based on the rules of the bitwiseAdjust function.
+
+    :raises ValueError: if the mask is not greater than 0.
     """
     if mask < 1:
         raise ValueError('Mask MUST be greater than 0')
@@ -93,6 +98,19 @@ def ceilDiv(n : int, d : int) -> int:
     outside the precision range of float.
     """
     return -(n // -d)
+
+def createZipOpen(func):
+    """
+    Creates a wrapper for the open function of a ZipFile that will automatically
+    set the current date as the modified time to the current time.
+    """
+    def _open(name, mode, *args, **kwargs):
+        if mode == 'w':
+            name = zipfile.ZipInfo(name, datetime.datetime.now().timetuple())
+
+        return func(name, mode, *args, **kwargs)
+
+    return _open
 
 def divide(string, length : int) -> list:
     """
@@ -123,8 +141,11 @@ def fromTimeStamp(stamp) -> datetime.datetime:
 def getCommandArgs(args):
     """
     Parse command-line arguments.
+
+    :raises IncompatibleOptionsError: if options are provided that are
+        incompatible.
     """
-    parser = argparse.ArgumentParser(description=constants.MAINDOC, prog='extract_msg')
+    parser = argparse.ArgumentParser(description = constants.MAINDOC, prog = 'extract_msg')
     # --use-content-id, --cid
     parser.add_argument('--use-content-id', '--cid', dest='cid', action='store_true',
                         help='Save attachments by their Content ID, if they have one. Useful when working with the HTML body.')
@@ -160,27 +181,33 @@ def getCommandArgs(args):
                         help='Tells the program to dump the message body (plain text) to stdout. Overrides saving arguments.')
     # --html
     parser.add_argument('--html', dest='html', action='store_true',
-                       help='Sets whether the output should be html. If this is not possible, will error.')
+                        help='Sets whether the output should be HTML. If this is not possible, will error.')
+    # --prepared-html
+    parser.add_argument('--prepared-html', dest='preparedHtml', action='store_true',
+                        help='When used in conjunction with --html, sets whether the HTML output should be prepared for embedded attachments.')
     # --raw
     parser.add_argument('--raw', dest='raw', action='store_true',
-                       help='Sets whether the output should be html. If this is not possible, will error.')
+                        help='Sets whether the output should be HTML. If this is not possible, will error.')
     # --rtf
     parser.add_argument('--rtf', dest='rtf', action='store_true',
-                       help='Sets whether the output should be rtf. If this is not possible, will error.')
+                        help='Sets whether the output should be RTF. If this is not possible, will error.')
     # --allow-fallback
     parser.add_argument('--allow-fallback', dest='allowFallback', action='store_true',
-                       help='Tells the program to fallback to a different save type if the selected one is not possible.')
+                        help='Tells the program to fallback to a different save type if the selected one is not possible.')
+    # --zip
+    parser.add_argument('--zip', dest='zip',
+                        help='Path to use for saving to a zip file.')
     # --out-name NAME
-    parser.add_argument('--out-name', dest = 'out_name',
-                        help = 'Name to be used with saving the file output. Should come immediately after the file name.')
+    parser.add_argument('--out-name', dest='out_name',
+                        help='Name to be used with saving the file output. Should come immediately after the file name.')
     # [msg files]
     parser.add_argument('msgs', metavar='msg', nargs='+',
-                        help='An msg file to be parsed')
+                        help='An msg file to be parsed.')
 
     options = parser.parse_args(args)
 
     # Check if more than one of the following arguments has been specified
-    if options.html + options.rtf + options.json > 1:
+    if options.html + options.rtf + options.json + options.raw > 1:
        raise IncompatibleOptionsError('Only one of these options may be selected at a time: --html, --json, --raw, --rtf')
 
     if options.dev or options.file_logging:
@@ -219,15 +246,12 @@ def getCommandArgs(args):
     options.msgs = file_tables
     return options
 
-def getContFileDir(_file_):
-    """
-    Takes in the path to a file and tries to return the containing folder.
-    """
-    return '/'.join(_file_.replace('\\', '/').split('/')[:-1])
-
 def getEncodingName(codepage : int) -> str:
     """
     Returns the name of the encoding with the specified codepage.
+
+    :raises UnknownCodepageError: if the codepage is unrecognized.
+    :raises UnsupportedEncodingError: if the codepage is not supported.
     """
     if codepage not in constants.CODE_PAGES:
         raise UnknownCodepageError(str(codepage))
@@ -246,13 +270,32 @@ def hasLen(obj) -> bool:
     """
     return hasattr(obj, '__len__')
 
-def injectHtmlHeader(msgFile) -> bytes:
+def injectHtmlHeader(msgFile, prepared : bool = False) -> bytes:
     """
     Returns the HTML body from the MSG file (will check that it has one) with
     the HTML header injected into it.
+
+    :param prepared: Determines whether to be using the standard HTML (False) or
+                     the prepared HTML (True) body (Default: False).
+
+    :raises AttributeError: if the correct HTML body cannot be acquired.
     """
     if not hasattr(msgFile, 'htmlBody') or not msgFile.htmlBody:
         raise AttributeError('Cannot inject the HTML header without an HTML body attribute.')
+
+    body = None
+
+    # We don't do this all at once because the prepared body is not cached.
+    if prepared:
+        if hasattr(msgFile, 'htmlBodyPrepared'):
+            body = msgFile.htmlBodyPrepared
+
+        # If the body is not valid or not found, raise an AttributeError.
+        if not body:
+            raise AttributeError('Cannot find a prepared HTML body to inject into.')
+
+    else:
+        body = msgFile.htmlBody
 
     def replace(bodyMarker):
         """
@@ -269,12 +312,15 @@ def injectHtmlHeader(msgFile) -> bytes:
         }).encode('utf-8')
 
     # Use the previously defined function to inject the HTML header.
-    return constants.RE_HTML_BODY_START.sub(replace, msgFile.htmlBody, 1)
+    return constants.RE_HTML_BODY_START.sub(replace, body, 1)
 
 def injectRtfHeader(msgFile) -> bytes:
     """
     Returns the RTF body from the MSG file (will check that it has one) with the
     RTF header injected into it.
+
+    :raises AttributeError: if the RTF body cannot be acquired.
+    :raises RuntimeError: if all injection attempts fail.
     """
     if not hasattr(msgFile, 'rtfBody') or not msgFile.rtfBody:
         raise AttributeError('Cannot inject the RTF header without an RTF body attribute.')
@@ -346,7 +392,7 @@ def injectRtfHeader(msgFile) -> bytes:
         logger.debug('Successfully injected RTF header using sixth method.')
         return data
 
-    raise Exception('All injection attempts failed.')
+    raise RuntimeError('All injection attempts failed. Please report this to the developer.')
 
 def inputToBytes(stringInputVar, encoding) -> bytes:
     """
@@ -506,7 +552,8 @@ def parseType(_type : int, stream, encoding, extras):
     For that example, extras should be a list of the bytes from rest of the
     streams.
 
-    WARNING: Not done. Do not try to implement anywhere where it is not already implemented
+    :raises NotImplementedError: for types with no current support. Most of
+        these types have no documentation of existing in an MSG file.
     """
     # WARNING Not done. Do not try to implement anywhere where it is not already implemented.
     value = stream
@@ -534,13 +581,16 @@ def parseType(_type : int, stream, encoding, extras):
     elif _type == 0x000A:  # PtypErrorCode
         value = constants.STUI32.unpack(value)[0]
         # TODO parsing for this.
-        # I can't actually find any msg properties that use this, so it should be okay to release this function without support for it.
-        raise NotImplementedError('Parsing for type 0x000A has not yet been implmented. If you need this type, please create a new issue labeled "NotImplementedError: parseType 0x000A"')
+        # I can't actually find any msg properties that use this, so it should
+        # be okay to release this function without support for it.
+        raise NotImplementedError('Parsing for type 0x000A (PtypErrorCode) has not yet been implmented. If you need this type, please create a new issue labeled "NotImplementedError: parseType 0x000A PtypErrorCode".')
     elif _type == 0x000B:  # PtypBoolean
         return constants.ST3.unpack(value)[0] == 1
     elif _type == 0x000D:  # PtypObject/PtypEmbeddedTable
         # TODO parsing for this
-        # Wait, that's the extension for an attachment folder, so parsing this might not be as easy as we would hope. The function may be released without support for this.
+        # Wait, that's the extension for an attachment folder, so parsing this
+        # might not be as easy as we would hope. The function may be released
+        # without support for this.
         raise NotImplementedError('Current version of extract-msg does not support the parsing of PtypObject/PtypEmbeddedTable in this function.')
     elif _type == 0x0014:  # PtypInteger64
         return constants.STI64.unpack(value)[0]
@@ -559,19 +609,24 @@ def parseType(_type : int, stream, encoding, extras):
     elif _type == 0x0048:  # PtypGuid
         return bytesToGuid(value)
     elif _type == 0x00FB:  # PtypServerId
-        # TODO parsing for this
-        raise NotImplementedError('Parsing for type 0x00FB has not yet been implmented. If you need this type, please create a new issue labeled "NotImplementedError: parseType 0x00FB"')
+        count = constants.STUI16.unpack(value[:2])
+        # If the first byte is a 1 then it uses the ServerID structure.
+        if value[3] == 1:
+            from .data import ServerID
+            return ServerID(value)
+        else:
+            return (count, value[2:count + 2])
     elif _type == 0x00FD:  # PtypRestriction
         # TODO parsing for this
-        raise NotImplementedError('Parsing for type 0x00FD has not yet been implmented. If you need this type, please create a new issue labeled "NotImplementedError: parseType 0x00FD"')
+        raise NotImplementedError('Parsing for type 0x00FD (PtypRestriction) has not yet been implmented. If you need this type, please create a new issue labeled "NotImplementedError: parseType 0x00FD PtypRestriction".')
     elif _type == 0x00FE:  # PtypRuleAction
         # TODO parsing for this
-        raise NotImplementedError('Parsing for type 0x00FE has not yet been implmented. If you need this type, please create a new issue labeled "NotImplementedError: parseType 0x00FE"')
+        raise NotImplementedError('Parsing for type 0x00FE (PtypRuleAction) has not yet been implmented. If you need this type, please create a new issue labeled "NotImplementedError: parseType 0x00FE PtypRuleAction".')
     elif _type == 0x0102:  # PtypBinary
         return value
     elif _type & 0x1000 == 0x1000:  # PtypMultiple
         # TODO parsing for `multiple` types
-        if _type in (0x101F, 0x101E):
+        if _type in (0x101F, 0x101E): # PtypMultipleString/PtypMultipleString8
             ret = [x.decode(encoding) for x in extras]
             lengths = struct.unpack(f'<{len(ret)}i', stream)
             lengthLengths = len(lengths)
@@ -581,7 +636,7 @@ def parseType(_type : int, stream, encoding, extras):
                 if lengths[x] != len(y):
                     logger.warning(f'Error while parsing multiple type. Expected length {lengths[x]}, got {len(y)}. Ignoring.')
             return ret
-        elif _type == 0x1102:
+        elif _type == 0x1102: # PtypMultipleBinary
             ret = copy.deepcopy(extras)
             lengths = tuple(constants.STUI32.unpack(stream[pos*8:(pos+1)*8])[0] for pos in range(len(stream) // 8))
             lengthLengths = len(lengths)
@@ -594,25 +649,25 @@ def parseType(_type : int, stream, encoding, extras):
         elif _type in (0x1002, 0x1003, 0x1004, 0x1005, 0x1007, 0x1014, 0x1040, 0x1048):
             if stream != len(extras):
                 logger.warning(f'Error while parsing multiple type. Expected {stream} entr{"y" if stream == 1 else "ies"}, got {len(extras)}. Ignoring.')
-            if _type == 0x1002:
+            if _type == 0x1002: # PtypMultipleInteger16
                 return tuple(constants.STMI16.unpack(x)[0] for x in extras)
-            if _type == 0x1003:
+            if _type == 0x1003: # PtypMultipleInteger32
                 return tuple(constants.STMI32.unpack(x)[0] for x in extras)
-            if _type == 0x1004:
+            if _type == 0x1004: # PtypMultipleFloating32
                 return tuple(constants.STMF32.unpack(x)[0] for x in extras)
-            if _type == 0x1005:
+            if _type == 0x1005: # PtypMultipleFloating64
                 return tuple(constants.STMF64.unpack(x)[0] for x in extras)
-            if _type == 0x1007:
+            if _type == 0x1007: # PtypMultipleFloatingTime
                 values = tuple(constants.STMF64.unpack(x)[0] for x in extras)
-                raise NotImplementedError('Parsing for type 0x1007 has not yet been implmented. If you need this type, please create a new issue labeled "NotImplementedError: parseType 0x1007"')
-            if _type == 0x1014:
+                return tuple(constants.PYTPFLOATINGTIME_START + datetime.timedelta(days = amount) for amount in values)
+            if _type == 0x1014: # PtypMultipleInteger64
                 return tuple(constants.STMI64.unpack(x)[0] for x in extras)
-            if _type == 0x1040:
+            if _type == 0x1040: # PtypMultipleTime
                 return tuple(filetimeToUtc(constants.ST3.unpack(x)[0]) for x in extras)
-            if _type == 0x1048:
+            if _type == 0x1048: # PtypMultipleGuid
                 return tuple(bytesToGuid(x) for x in extras)
         else:
-            raise NotImplementedError(f'Parsing for type {_type} has not yet been implmented. If you need this type, please create a new issue labeled "NotImplementedError: parseType {_type}"')
+            raise NotImplementedError(f'Parsing for type {_type} has not yet been implmented. If you need this type, please create a new issue labeled "NotImplementedError: parseType {_type}".')
     return value
 
 def prepareFilename(filename) -> str:
@@ -697,56 +752,57 @@ def rtfSanitizePlain(inp : str) -> str:
 
     return output
 
-def setupLogging(defaultPath=None, defaultLevel=logging.WARN, logfile=None, enableFileLogging=False,
-                  env_key='EXTRACT_MSG_LOG_CFG'):
+def setupLogging(defaultPath = None, defaultLevel = logging.WARN, logfile = None, enableFileLogging : bool = False,
+                  env_key = 'EXTRACT_MSG_LOG_CFG') -> bool:
     """
     Setup logging configuration
 
     Args:
-        defaultPath (str): Default path to use for the logging configuration file
-        defaultLevel (int): Default logging level
-        env_key (str): Environment variable name to search for, for setting logfile path
+    :param defaultPath: Default path to use for the logging configuration file.
+    :param defaultLevel: Default logging level.
+    :param env_key: Environment variable name to search for, for setting logfile
+        path.
+    :param enableFileLogging: Whether to use a file to log or not.
 
     Returns:
         bool: True if the configuration file was found and applied, False otherwise
     """
-    shippedConfig = getContFileDir(__file__) + '/logging-config/'
+    shippedConfig = pathlib.Path(__file__).parent / 'logging-config'
     if os.name == 'nt':
         null = 'NUL'
-        shippedConfig += 'logging-nt.json'
+        shippedConfig /= 'logging-nt.json'
     elif os.name == 'posix':
         null = '/dev/null'
-        shippedConfig += 'logging-posix.json'
+        shippedConfig /= 'logging-posix.json'
     # Find logging.json if not provided
-    if not defaultPath:
-        defaultPath = shippedConfig
+    defaultPath = pathlib.Path(defaultPath) if defaultPath else shippedConfig
 
     paths = [
         defaultPath,
-        'logging.json',
-        '../logging.json',
-        '../../logging.json',
+        pathlib.Path('logging.json'),
+        pathlib.Path('../logging.json'),
+        pathlib.Path('../../logging.json'),
         shippedConfig,
     ]
 
     path = None
 
     for configPath in paths:
-        if os.path.exists(configPath):
+        if configPath.exists():
             path = configPath
             break
 
     value = os.getenv(env_key, None)
-    if value and os.path.exists(value):
-        path = value
+    if value and os.path.exists(value) and os.path.isfile(value):
+        path = pathlib.Path(value)
 
-    if path is None:
+    if not path:
         print('Unable to find logging.json configuration file')
         print('Make sure a valid logging configuration file is referenced in the defaultPath'
               ' argument, is inside the extract_msg install location, or is available at one '
               'of the following file-paths:')
         print(str(paths[1:]))
-        logging.basicConfig(level=defaultLevel)
+        logging.basicConfig(level = defaultLevel)
         logging.warning('The extract_msg logging configuration was not found - using a basic configuration.'
                         f'Please check the extract_msg installation directory for "logging-{os.name}.json".')
         return False
@@ -759,8 +815,8 @@ def setupLogging(defaultPath=None, defaultLevel=logging.WARN, logfile=None, enab
             if enableFileLogging:
                 config['handlers'][x]['filename'] = tmp = os.path.expanduser(
                     os.path.expandvars(logfile if logfile else config['handlers'][x]['filename']))
-                tmp = getContFileDir(tmp)
-                if not os.path.exists(tmp):
+                tmp = pathlib.Path(tmp).parent
+                if not tmp.exists:
                     os.makedirs(tmp)
             else:
                 config['handlers'][x]['filename'] = null
@@ -774,10 +830,14 @@ def setupLogging(defaultPath=None, defaultLevel=logging.WARN, logfile=None, enab
     logging.getLogger().setLevel(defaultLevel)
     return True
 
-def verifyPropertyId(id):
+def verifyPropertyId(id : str) -> None:
     """
-    Determines whether a property ID is valid for the functions that this function
-    is called from. Property IDs MUST be a 4 digit hexadecimal string.
+    Determines whether a property ID is valid for vertain functions. Property
+    IDs MUST be a 4 digit hexadecimal string. Property is valid if no exception
+    is raised.
+
+    :raises InvaildPropertyIdError: if the it is not a 4 digit hexadecimal
+        number.
     """
     if not isinstance(id, str):
         raise InvaildPropertyIdError('ID was not a 4 digit hexadecimal string')
@@ -790,9 +850,14 @@ def verifyPropertyId(id):
             raise InvaildPropertyIdError('ID was not a 4 digit hexadecimal string')
 
 def verifyType(_type):
+    """
+    Verifies that the type is valid. Raises an exception if it is not.
+
+    :raises UnknownTypeError: if the type is not recognized.
+    """
     if _type is not None:
         if (_type not in constants.VARIABLE_LENGTH_PROPS_STRING) and (_type not in constants.FIXED_LENGTH_PROPS_STRING):
-            raise UnknownTypeError(f'Unknown type {_type}')
+            raise UnknownTypeError(f'Unknown type {_type}.')
 
 def windowsUnicode(string):
     return str(string, 'utf-16-le') if string is not None else None